skillverify-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.0
pydantic==2.6.1
pydantic-settings==2.2.1
httpx==0.27.0
playwright==1.42.0
firebase-admin==6.4.0
google-generativeai==0.4.1
python-multipart==0.0.9
